requirements.txt
flask
python-dotenv
huggingface-hub
cohere
PyPDF2
